Stop timing the harness, and stop the runtime tier re-transcoding its own static text
Step 0 of the backend-profiling plan: the assessment findings that sit directly on
the measurement path, landed before any profiling so the profiler measures the
engine rather than the instrument. ASSESSMENT-FINDINGS.md is added as the record
those fixes answer to.

## The sink rows were mostly measuring FNV-1a

The gate's checksum writers fold a hash over every unit of output -- the
TextWriter one through a per-char virtual call -- and allocate a fresh 64 KB
buffer per render. That is right for a gate, which must prove the engine produced
the bytes, and wrong for a bench, which must measure the engine. No competitor row
paid anything comparable.

Bench/BenchSinks.cs adds the two dumbest honest sinks: a byte writer that advances
an offset and nothing else, and a char writer that does one bulk CopyTo per write.
Both are pre-sized past their high-water mark in [GlobalSetup] and reused with
Reset(), so no allocation and no resize lands in the timed region. RenderToSink is
replaced by RenderToBuffer/RenderToWriter/RenderToString, which take the model and
the sink from the caller.

Dropping the per-iteration checksum is safe because the property it establishes --
the engine really produced every unit -- is a per-process fact. It is proven,
untimed, by the gate, by SelfTest's differential, and now by an assertion in each
suite's [GlobalSetup] that both bench sinks reproduce the gated string render. The
checksum writers are untouched on the gate path.

Measured, runtime utf8 technique row, before -> after:

  composed-page      41,775 ns ->  1,930 ns   (allocated 65,808 B -> 184 B)
  encoded-loop    2,876,043 ns -> 719,822 ns
  large-loop        623,079 ns -> 156,397 ns
  fragment-heavy      7,062 ns ->   3,438 ns
  fortunes-encoded    2,216 ns ->     699 ns

The overhead scaled with output size, as a per-byte fold would. On composed-page
the old cell was ~95% harness.

Taking the model as a parameter also closes an asymmetry that was scheduled for
later: TechniqueRuntimeBenchmarks called ModelFor inside every timed iteration
while TechniquePrecompiledBenchmarks hoisted it in setup, so the runtime side was
paying for a model allocation its rival never did.

## The runtime tier re-encoded static text on every UTF-8 render

RuntimeDocument's static pieces are UTF-16 strings, so every UTF-8 render called
Encoding.UTF8.GetBytes over all static content again; the precompiled tier has had
pre-encoded u8 literals through WritePiece all along. DataProcessor now carries a
PieceUtf8 cache built once at document build, and both render sites branch on
IUtf8ScopeRenderer in exactly WritePiece's shape. Static pieces reach the sink
directly and never pass the encode proxy, so the never-bypass-encoding invariant
is untouched.

This was the whole of the generated backend's only win. On the utf8 sink, after
both fixes, generated is now slower everywhere:

  trivial-substitution   132.8 ns vs   201.8 ns   1.52x slower
  mixed-page           2,744.6 ns vs 4,423.0 ns   1.61x slower
  fragment-heavy       3,438.4 ns vs 4,225.8 ns   1.23x slower
  conditional-heavy   18,307.7 ns vs 24,573.2 ns  1.34x slower
  large-loop         156,397.1 ns vs 218,228.0 ns 1.40x slower  (was a 1.11x WIN)

Directional only: these are single-launch --job Short runs, not the protocol
budget. The published tables are re-measured separately.

## Name collision the sink flip created

HeddleEngine.Name was still "Heddle (utf8)" after the anchor moved to the string
sink, so it both mislabelled the anchor and computed the same label as the real
utf8 cell -- two indistinguishable gate lines per workload, and any lookup keyed
on Cell.Engine alone ambiguous. Renamed to "Heddle"; the technique cells are now
"Heddle (utf8 sink)" and "Heddle (textwriter sink)", matching the report's own
vocabulary. consolidate.py keys .NET rows off benchmark method names, so the
report is unaffected.

Also corrected: HeddleEngine's class header and Registry's InCrossStack doc, both
still arguing the pre-flip utf8-anchor rationale the current code repudiates; a
dangling cref left by an earlier method split; and ChecksumTextWriter reporting
Encoding.UTF8 while consuming chars.

## Verification

gate 143/143, selftest 141/141, verify-corpus clean, and gate-precompiled still
byte-for-byte across all 5 workloads x 3 sinks. Full suite 2134 tests with 4
failures, all pre-existing and none from this change: the docs/benchmarks links
awaiting the new report's index.md, two Windows path-separator expectations, and
test-written artifacts in the shared corpus directory. The public-API golden gains
HED7031's new const, which is the one failure this work did cause.

Deferred: the HtmlEncodedRenderer allocation fix (finding 4). Its clean-span fast
path needs FindFirstCharacterToEncode, whose char-span form requires unsafe
pointers, so it is a TFM-conditional change across netstandard2.0/net8/net10 --
and it cannot affect the generated-vs-runtime question, because neither encoded
workload has a precompiled counterpart.

+++ benchmarks/dotnet/src/Bench/CrossStackBenchmarks.cs
@@ -19,8 +19,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "composed-page";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
@@ -34,8 +34,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "trivial-substitution";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
@@ -49,8 +49,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "large-loop";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
@@ -64,8 +64,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "mixed-page";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
@@ -79,8 +79,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "conditional-heavy";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
@@ -94,8 +94,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "fragment-heavy";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
@@ -109,8 +109,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "fortunes-encoded";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
@@ -124,8 +124,8 @@ namespace Heddle.Benchmarks.Dotnet.Bench
         protected override string Workload => "encoded-loop";
 
         [Benchmark(Baseline = true)] public string RenderHeddle() => Render("Heddle");
-        [Benchmark] public ulong RenderHeddleUtf8() => RenderHeddleUtf8Sink();
-        [Benchmark] public ulong RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
+        [Benchmark] public int RenderHeddleUtf8() => RenderHeddleUtf8Sink();
+        [Benchmark] public int RenderHeddleTextWriter() => RenderHeddleTextWriterSink();
         [Benchmark] public string RenderFluid() => Render("Fluid");
         [Benchmark] public string RenderScriban() => Render("Scriban");
         [Benchmark] public string RenderDotLiquid() => Render("DotLiquid");
